Remove shortcode, add button & dialog by hooking into actions
