+++ plugins/wporg-5ftf/views/button-request-manage-link.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace WordPressDotOrg\FiveForTheFuture\View;
+
+defined( 'WPINC' ) || die();
+
+?>
+<div class="edit-pledge-wrapper">
+	<button id="toggle-management-link-form" class="button button-link">
+		<span class="dashicons dashicons-edit" aria-hidden="true"></span>
+		<?php esc_html_e( 'Edit Pledge', 'wporg-5ftf' ); ?>
+	</button>
+</div>
